Script to convert 'Redirecting for..' JSON to donation queue message
We currently log the provisional donation queue message prior to
redirecting donors to gateway forms. The JSON message is output
next to the string 'Redirecting for transaction:' and can be
extracted and fed into this script to along with a real
gateway_txn_id to generate a synthetic donation queue message.

To test:
./redirectJson2DonationJson.sh "NEW_TXN_ID_HERE" '{"gateway_txn_id":false,"response":false,"gateway_account":"WikimediaDonations","fee":0,"contribution_tracking_id":"3","utm_source":"..cc","language":"en","email":"jwales@example.com","first_name":"Jimmy","last_name":"Wales","street_address":"1 Montgomery Street","city":"San Francisco","state_province":"CA","country":"US","postal_code":"94104","gateway":"adyen","order_id":"3.1","recurring":"","payment_method":"cc","payment_submethod":"","currency":"USD","gross":"10.00","user_ip":"172.18.0.1","date":1694542101}'